+++ src/lib/mocks.ts
@@ -3,7 +3,7 @@ export const feedItemsMocked = [
     id: 1,
     imgUrl:
       "https://png.pngtree.com/thumb_back/fw800/background/20190221/ourmid/pngtree-hand-painted-beautiful-outdoor-outskirts-image_32488.jpg",
-    description: `Lorem ipsum dolor sit, amet consectetur adipisicing elit. Ducimus quo
+    message: `Lorem ipsum dolor sit, amet consectetur adipisicing elit. Ducimus quo
               explicabo qui voluptates, eaque distinctio dolor, ab facilis sequi
               eius aliquid recusandae consequuntur hic voluptas deserunt`,
     name: "Manuela Oliveira",
@@ -12,14 +12,14 @@ export const feedItemsMocked = [
     id: 2,
     imgUrl:
       "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTERXnmbI1gv3m0GaiWvVTQ1PpxyykTASgOYA&s",
-    description: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam ut nisi justo. Donec scelerisque dui non magna viverra, sit amet feugiat leo vehicula.`,
+    message: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam ut nisi justo. Donec scelerisque dui non magna viverra, sit amet feugiat leo vehicula.`,
     name: "João Silva",
   },
   {
     id: 3,
     imgUrl:
       "https://i.pinimg.com/236x/59/19/26/591926a3eb106ee1851a1e574fec0edb.jpg",
-    description: `Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Duis laoreet quam eget neque bibendum, sit amet blandit justo tristique.`,
+    message: `Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Duis laoreet quam eget neque bibendum, sit amet blandit justo tristique.`,
     name: "Maria Souza",
   },
 ];
